RacingCarGame의 UI관련 로직을 Application으로 이동하여 책임 분리
